client: move V1Client.serverHost code to its better position
