+++ tools/client/clients/ai/client.go
@@ -12,7 +12,9 @@ import (
 	"errors"
 	"fmt"
 	"io"
+	"log"
 	"net/http"
+	"net/url"
 	"runtime/debug"
 	"strconv"
 
@@ -108,6 +110,14 @@ func (c *V1Client) UpgradeOptional() (neo Client, err error) {
 
 	neo = NewV2Client(c.endpoint, k, token.UserID)
 	return neo, nil
+}
+
+func (c *V1Client) serverHost() string {
+	u, err := url.Parse(c.endpoint)
+	if err != nil {
+		log.Fatal(err)
+	}
+	return fmt.Sprintf("%s://%s", u.Scheme, u.Host)
 }
 
 func doRequestAndVerifyStatusReadBodyAll(req *http.Request) (responsePayload []byte, err error) {
